Bootstrap github actions
This bootstraps a 'test' workflow that sets up a go project and runs
`make test`.

caching is disabled in the go setup stage, this is due to it fumbling
the `go.work` sum file, that is currently empty. It's probably possible
to re-enable it later when there is actual code running.

It's nice to see that `make test` works without issue on windows.
Ideally this all will, but it's likely that the windows test machine
will be a great canary for my mac/linux based development workflow.