Add support for running benchmarks with `sweet.Run`
`sweet.Run` now can run benchmarks with the same interface as for
tests. Pass in a `*testing.B` where you would otherwise pass a
`*testing.T` and it'll work exactly the same.

It doesn't yet make any special allowances for benchmarks. In particular
it doesn't call `ResetTimer` after creating it's dependencies. If this
is desired it can be handled in your factory function.