Turn factory/testcontainer into it's own pacakge
By moving the package up a level we can leverage the same interface for
creating and managing containers for test.

This should allow us centralise the test waiting logic and cleanup and
allow compatiblility for docker-compose based tests. While also making
it easier to bootstrap other test deps.

The package has been renamed `sweet/factories/tc` so we don't conflict
with the `testcontainers` package. Mostly in tests, but the packages are
likely to show up together in use cases too.