Fix build
All go packages support go 1.18, and the build builds from there up.

Golangci lint was failing due to an esoteric bug, so it's been
upgraded.

Windows fails on this test due to known docker issues; disabling on
windows until we can fix it.

The toolchain line was removed to avoid issues with golangci-lint.
Revisit when it's updated or with a better linter.

A longer timeout was given to golangci-lint for windows builds. See
revisiting with a better linter...