Clarify modes of use in the package docs
This goes into the various ways of defining dependencies, their pros and
cons.

The examples are also defined in `example_test.go` which will add them
to the go docs and verify that they compile properly.

The package documentation and readme have been aligned. The focus is on
the package docs where links to packages and functions work and examples
are compiled and linked up.

The readme focuses on the bigger pitch, overview and idea of the project.